tiny change

+++ InherandPoly/src/main/java/vehicles/Car.java
@@ -17,7 +17,7 @@ public class Car extends Vehicle {
             default -> throw new IllegalStateException("Unexpected value: " + this.engine.fuelType);
         };
         for (int i = 0 ; i < 10 ; i++) {
-            System.out.println(message + String.format("%.2f", (engine.horsepower
+            System.out.println(message + String.format("%.2f", ((int)engine.horsepower
                     * fuelMod) * Math.exp(Math.sqrt(i)) / 2));
         }
