+++ InherandPoly/src/main/java/vehicles/Motorcycle.java
@@ -19,8 +19,8 @@ public class Motorcycle extends Vehicle{
             default -> throw new IllegalStateException("Unexpected value: " + this.engine.fuelType);
         };
         for (int i = 0 ; i < 10 ; i++) {
-            System.out.println(message + String.format("%.2f", ((engine.horsepower * fuelMod)
-                    * Math.exp(Math.sqrt(i)) / 2)* (hasSidecar ? 0.8 : 1.0)));
+            System.out.println(message + String.format("%.2f", (((int)engine.horsepower * fuelMod)
+                    * Math.exp(Math.sqrt(i)) / 2)* ((this.hasSideCar ? 0.8 :  1.0))));
         }
 
     }
